ensured http verbs and added url parameters

+++ hangouts/views.py
@@ -8,6 +8,9 @@ from django.core import serializers
 
 def create_hangout(request):
   try:
+    #ensure correct http verb
+    if request.method != "POST":
+      return HttpResponse("This is a POST endpoint", status=405)
     #get json palyload
     json_payload = json.loads(request.body)
     # create new entry
@@ -29,10 +32,13 @@ def create_hangout(request):
     return HttpResponse("Entry added successfull")
   except Exception as e: 
     print("Error Occured:", e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 def read_all(request):
   try:
+    #ensure correct http verb
+    if request.method != "GET":
+      return HttpResponse("This is a GET endpoint", status=405)
     # get all objects
     all_hangouts = HangoutEntry.objects.all()
     # convert to json
@@ -41,31 +47,32 @@ def read_all(request):
     return HttpResponse(returning_json, content_type="application/json")
   except Exception as e:
     print("Error Occured:", e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 
-def read_one(request):
+def read_one(request, pk):
   try:
-    # read request payload
-    json_payload = json.loads(request.body)
+    #ensure correct http verb
+    if request.method != "GET":
+      return HttpResponse("This is a GET endpoint", status=405)
     # get object
-    hangout = HangoutEntry.objects.get(pk=json_payload['pk'])
+    hangout = HangoutEntry.objects.get(pk=pk)
     hangout_serialized = serializers.serialize('json', [hangout])
-    print('hangout', hangout_serialized)
     return HttpResponse(hangout_serialized, content_type="application/json")
   except Exception as e:
     print('error:', e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 
-def update_one(request):
+def update_one(request, pk):
   try:
+    #ensure correct http verb
+    if request.method != "PUT":
+      return HttpResponse("This is a PUT endpoint", status=405)
     # read payload
-    json_payload = json.loads(request.body)
-    requested_pk = json_payload['editingPk']
-    new_parameters = json_payload['newParameters']
+    new_parameters = json.loads(request.body)
     # get object reference
-    hangout_to_edit = HangoutEntry.objects.get(pk=requested_pk)
+    hangout_to_edit = HangoutEntry.objects.get(pk=pk)
     # edit object reference
     hangout_to_edit.name = new_parameters['name']
     hangout_to_edit.location = new_parameters['location']
@@ -82,21 +89,22 @@ def update_one(request):
     return HttpResponse("Entry updated")
   except Exception as e:
     print('error:', e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
-def delete_one(request):
+def delete_one(request, pk):
   try:
-    # read request payload
-    json_payload = json.loads(request.body)
+    #ensure correct http verb
+    if request.method != "DELETE":
+      return HttpResponse("This is a DELETE endpoint", status=405)
     # get object
-    hangout = HangoutEntry.objects.get(pk=json_payload['pk'])
+    hangout = HangoutEntry.objects.get(pk=pk)
     # delete
     hangout.delete()
     # return all a-okay
     return HttpResponse("deleted")
   except Exception as e:
     print('error:', e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 
 
@@ -104,6 +112,9 @@ def delete_one(request):
 ## ---------- only for development
 
 def dev_delete_all(request):
+  #ensure correct http verb
+  if request.method != "DELETE":
+    return HttpResponse("This is a DELETE endpoint", status=405)
   # get all objects
   all_hangouts_dev = HangoutEntryDev.objects.all()
   # for all objects delete
@@ -113,6 +124,9 @@ def dev_delete_all(request):
 
 def dev_create_dummy_dataset(request):
   try:
+    #ensure correct http verb
+    if request.method != "POST":
+      return HttpResponse("This is a POST endpoint", status=405)
     #create dummy data
     for data in dummy_data:
       devEntry = HangoutEntryDev(
@@ -130,11 +144,14 @@ def dev_create_dummy_dataset(request):
       devEntry.save()
     return HttpResponse("dummy dev data created")
   except Exception as e:
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 
 def dev_create_hangout(request):
   try:
+    #ensure correct http verb
+    if request.method != "POST":
+      return HttpResponse("This is a POST endpoint", status=405)
     #get json palyload
     json_payload = json.loads(request.body)
     # create new entry
@@ -156,10 +173,13 @@ def dev_create_hangout(request):
     return HttpResponse("dev Entry added successfull")
   except Exception as e: 
     print("Error Occured:", e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 def dev_read_all(request):
   try:
+    #ensure correct http verb
+    if request.method != "GET":
+      return HttpResponse("This is a GET endpoint", status=405)
     # get all objects
     all_hangouts = HangoutEntryDev.objects.all()
     # convert to json
@@ -168,31 +188,32 @@ def dev_read_all(request):
     return HttpResponse(returning_json, content_type="application/json")
   except Exception as e:
     print("Error Occured:", e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 
-def dev_read_one(request):
+def dev_read_one(request, pk):
   try:
-    # read request payload
-    json_payload = json.loads(request.body)
+    #ensure correct http verb
+    if request.method != "GET":
+      return HttpResponse("This is a GET endpoint", status=405)
     # get object
-    hangout = HangoutEntryDev.objects.get(pk=json_payload['pk'])
+    hangout = HangoutEntryDev.objects.get(pk=pk)
     hangout_serialized = serializers.serialize('json', [hangout])
-    print('hangout', hangout_serialized)
     return HttpResponse(hangout_serialized, content_type="application/json")
   except Exception as e:
     print('error:', e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
 
-def dev_update_one(request):
+def dev_update_one(request, pk):
   try:
+    #ensure correct http verb
+    if request.method != "PUT":
+      return HttpResponse("This is a PUT endpoint", status=405)
     # read payload
-    json_payload = json.loads(request.body)
-    requested_pk = json_payload['editingPk']
-    new_parameters = json_payload['newParameters']
+    new_parameters = json.loads(request.body)
     # get object reference
-    hangout_to_edit = HangoutEntryDev.objects.get(pk=requested_pk)
+    hangout_to_edit = HangoutEntryDev.objects.get(pk=pk)
     # edit object reference
     hangout_to_edit.name = new_parameters['name']
     hangout_to_edit.location = new_parameters['location']
@@ -209,18 +230,19 @@ def dev_update_one(request):
     return HttpResponse("dev entry updated")
   except Exception as e:
     print('error:', e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
 
-def dev_delete_one(request):
+def dev_delete_one(request, pk):
   try:
-    # read request payload
-    json_payload = json.loads(request.body)
+    #ensure correct http verb
+    if request.method != "DELETE":
+      return HttpResponse("This is a DELETE endpoint", status=405)
     # get object
-    hangout = HangoutEntryDev.objects.get(pk=json_payload['pk'])
+    hangout = HangoutEntryDev.objects.get(pk=pk)
     # delete
     hangout.delete()
     # return all a-okay
     return HttpResponse("dev entry deleted")
   except Exception as e:
     print('error:', e)
-    return HttpResponse(e)
+    return HttpResponse(e, status=500)
